docs: Update README to reflect AIAgent aggregate details
The README is updated to include a description of the `AIAgent` aggregate, outlining its properties, functionalities, and relationship with the `User` aggregate.
The directory structure in the Readme has been updated with a comment to indicate the role of Domain directory.
The progress checklist has been updated to indicate that the AIAgent and Node aggregates have been completed.