feat: Add MessageException and MessageCouldNotBeEmpty exceptions
This commit introduces two new exception classes: MessageException and MessageCouldNotBeEmpty. These exceptions are used to handle scenarios where a message is not found or when a message is empty, respectively.

The MessageCouldNotBeEmpty exception is also added to the __init__.py file to make it accessible from the domain.exceptions package.