feat: Add EmailNotValid exception
This commit introduces a new exception, `EmailNotValid`, to the `domain.exceptions` module. This exception is raised when an email address is determined to be invalid according to domain rules.

The exception is also added to the `__init__.py` file to be accessible as part of the public API of the module.