feat: Add MessagesListNotValid exception
This commit introduces a new exception, `MessagesListNotValid`, to the domain exceptions module. This exception is raised when a list of messages is found to be invalid, allowing for more specific error handling in scenarios dealing with collections of messages.
The exception is also added to the `__init__.py` file to make it easily accessible.