refactor: Refactor ChatContext to use MessageCollection
This change introduces the `MessageCollection` object to manage the messages within a `ChatContext`.  Instead of storing messages in a dictionary, it now uses a dedicated `MessageCollection` object.  This change improves code organization and allows for future enhancements to message management. Removes `total_tokens` as this will be managed by the `MessageCollection`.