feat: Allow overriding limits in ChatContext.from_dict
This change allows overriding the `limits` attribute when creating a `ChatContext` object from a dictionary using the `from_dict` method. Previously, the `limits` were not considered and were being initialized with the object's default.

This allows more flexibility when reconstructing `ChatContext` objects from persisted data or configurations where the limits may need to be dynamically adjusted.