refactor: Simplify ChatContext message collection update
The `limits` parameter was removed from the `MessageCollection` constructor when updating a message within the `ChatContext`. This was redundant since the limits are a property of the context and don't need to be passed during updates.

The `total_tokens` attribute was removed from the clear_history method, because the chat context should only care about the messages.