refactor: Delegate message expiration logic to a service
This change refactors the `MessageCollection` class to delegate the message expiration logic to a `MessageExpirationService`.

This improves the separation of concerns and makes the `MessageCollection` class more focused on its core responsibility of managing messages. The expiration logic is now encapsulated in a dedicated service, which can be easily tested and maintained.