refactor: Use service instance for update_message in tests
The test `test_update_message_empty_collection` was directly calling the static method `MessageOperationsService.update_message` instead of using the service instance created by the fixture.

This commit refactors the test to use the `service` fixture instance to call the `update_message` method. This makes the test more consistent with other tests and aligns with how the service is intended to be used.